Created a separate dummyData file
