+++ FrontEnd/src/components/dummyServices.js
@@ -1,0 +1,20 @@
+const initialCategories = [
+    "Automotive Repair",
+    "Healthcare",
+    "Plumbing",
+    "Electrical",
+    "Legal Services",
+    "Beauty and Wellness",
+    "Home Maintenance",
+    "Financial Services",
+    "Babysitting",
+    "Furniture Repair",
+    "Gardening",
+    "Cleaning",
+    "Pest Control Services",
+    "Counseling",
+    "Managing",
+    "others"
+  ];
+  
+  export default initialCategories;
